create payment option updated in main menu
